requirements.txt
numpy
pandas
opencv-python
matplotlib
scipy
pyyaml
tqdm
torch
streamlit
ultralytics
pytest
